requirements.txt
absl-py==1.0.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.12
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
autokeras==1.0.20
Babel==2.9.1
bitsandbytes==0.39.0
black==23.3.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
clip-anytorch==2.5.2
cloudpickle==2.0.0
codecov==2.1.12
community==1.0.0b1
conda==4.12.0
conda-build==3.21.8
coverage==6.3.2
cuda-python==11.6.1
cycler==0.11.0
Cython==0.29.28
datasets==2.12.0
debugpy==1.5.1
defusedxml==0.7.1
diffusers==0.21.4
dill==0.3.6
dm-tree==0.1.8
docutils==0.17.1
entrypoints==0.4
expecttest==0.1.3
fairscale==0.4.13
fastrlock==0.8
fire==0.5.0
Flask==2.0.3
flatbuffers==1.12
fonttools==4.31.2
frozenlist==1.3.3
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gensim==4.3.2
glob2==0.7
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
HeapDict==1.0.1
huggingface-hub==0.14.1
hypothesis==4.50.8
ImageHash==4.3.1
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.2
ipython-genutils==0.2.0
itsdangerous==2.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab==2.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.13.7
keras==2.9.0
keras-core==0.1.5
Keras-Preprocessing==1.1.2
keras-tuner==1.4.5
kiwisolver==1.4.0
kt-legacy==1.0.5
libclang==16.0.6
librosa==0.9.1
lightning-utilities==0.8.0
llvmlite==0.36.0
lmdb==1.3.0
locket==0.2.1
Markdown==3.3.6
markdown-it-py==3.0.0
matplotlib==3.5.1
mdit-py-plugins==0.3.0
mdurl==0.1.2
mistune==0.8.4
msgpack==1.0.3
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
namex==0.0.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.1
nvidia-dali-cuda110==1.11.1
nvidia-pyindex==1.0.9
nvtx==0.2.4
oauthlib==3.2.0
open-clip-torch==2.23.0
opencv-contrib-python==4.5.5.62
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
partd==1.2.0
pathspec==0.11.1
Pillow==10.1.0
platformdirs==3.5.1
pluggy==1.0.0
polygraphy==0.33.0
pooch==1.6.0
prettytable==3.2.0
prometheus-client==0.13.1
protobuf==3.19.6
py==1.11.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.1
pydot==1.4.2
Pygments==2.16.1
pynvml==11.4.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
python-hostlist==1.21
python-louvain==0.16
python-nvd3==0.15.0
python-slugify==6.1.1
pytorch-lightning==2.0.2
pytorch-quantization==2.1.2
PyWavelets==1.4.1
pyzmq==22.3.0
regex==2022.3.15
requests-oauthlib==1.3.1
resampy==0.2.2
responses==0.18.0
rich==13.6.0
rsa==4.8
sacremoses==0.0.49
safetensors==0.3.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.99
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
Sphinx==4.4.0
sphinx-glpi-theme==0.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
timm==0.9.1
tokenize-rt==5.0.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.12.0a0+2c916ef
torchmetrics==0.11.4
torchtext==0.12.0a0
tornado==6.1
tqdm==4.63.0
transformers==4.29.1
webencodings==0.5.1
Werkzeug==2.0.3
wget==3.2
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zict==2.1.0
zipp==3.7.0
